# src/main/resources/static/openapi/swagger-v1.0.0.yaml
swagger: "2.0"
info:
  description: "This project is fake REST API for testing, and prototyping."
  version: "1.0.0"
  title: "Mokapi"
  contact:
    email: "burakkggul@gmail.com"
  license:
    name: "GNUv3"
    url: "https://github.com/burakkggul/mokapi/blob/master/LICENSE"
host: "localhost:8080"
basePath: "/"
...
schemes:
  - "https"
  - "http"
paths:
...
  /users/todos/{TodoID}:

    get:
      tags:
        - "Todos"
      summary: "Get todo by todo id."
      operationId: "getTodoByTodoID"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Access-Key"
          type: "string"
          format: string
          required: false
          description: "Access key providing data access authorization."
        - in: "path"
          name: "TodoID"
          required: true
          type: "string"
          format: "TodoID"
          description: "The ID of a specific todo"
      responses:
        "200":
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Todo"
        "404":
          description: "message:User not found. You can use the post method to add users."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"
        "500":
          description: "message: An error occurred. We took note to fix it as soon as possible."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"

    put:
      tags:
        - "Todos"
      summary: "Update an existing todo by todo id."
      operationId: "updateTodoByTodoId"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Access-Key"
          type: "string"
          format: string
          required: false
          description: "Access key providing data access authorization."
        - in: "path"
          name: "TodoID"
          required: true
          type: "string"
          format: "TodoID"
          description: "The ID of a specific todo"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Post"
      responses:
        "200":
          description: "Updated successfully."
          schema:
            $ref: "#/definitions/TodoInfoResponse"
        "404":
          description: "message:User not found"
          schema:
            $ref: "#/definitions/ErrorInfoResponse"
        "500":
          description: "message: An error occurred. We took note to fix it as soon as possible."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"

    patch:
      tags:
        - "Todos"
      summary: "Changes specific data of the todo by todo id. JSON Patch is specified in RFC 6902 from the IETF."
      operationId: "patchTodoByTodoId"
      consumes:
        - "application/json-patch+json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Access-Key"
          type: "string"
          format: string
          required: false
          description: "Access key providing data access authorization."
        - in: "path"
          name: "TodoID"
          required: true
          type: "string"
          format: "TodoID"
          description: "The ID of a specific todo"
        - name: "JsonPatch"
          in: "body"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/JsonPatch"
      responses:
        '200':
          description: "Patching was successful."
          schema:
            $ref: "#/definitions/TodoInfoResponse"
        "400":
          description: "message:Bad Request"
          schema:
            $ref: "#/definitions/InfoResponse"
        "404":
          description: "message:User not found. You can use the post method to add users."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"
        "500":
          description: "message: An error occurred. We took note to fix it as soon as possible."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"

    delete:
      tags:
        - "Todos"
      summary: "Delete todo by todo id."
      operationId: "deleteTodoByTodoId"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Access-Key"
          type: "string"
          format: string
          required: false
          description: "Access key providing data access authorization."
        - in: "path"
          name: "TodoID"
          required: true
          type: "string"
          format: "TodoID"
          description: "The ID of a specific todo"
      responses:
        "200":
          description: "Post has been successfully deleted."
          schema:
            $ref: "#/definitions/InfoResponse"

        "500":
          description: "message: An error occurred. We took note to fix it as soon as possible."
          schema:
            $ref: "#/definitions/ErrorInfoResponse"
...
definitions:
...
  Todo:
    type: "object"
    properties:
      title:
        type: "string"
      completed:
        type: "boolean"
...
  Post:
    type: "object"
    properties:
      title:
        type: "string"
      body:
        type: "string"
      comments:
        type: "array"
        items:
          $ref: "#/definitions/Comment"
  Comment:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      body:
        type: "string"
  InfoResponse:
    type: "object"
    properties:
      timestamp:
        type: "string"
        example: "2021-01-11T17:22:40.856+00:00"
      status:
        type: "integer"
        format: "int64"
        example: 200
      message:
        type: "string"
        example: "Success"
...
  TodoInfoResponse:
    type: "object"
    properties:
      timestamp:
        type: "string"
      status:
        type: "integer"
        format: "int64"
      message:
        type: "string"
      data:
        $ref: "#/definitions/Todo"
...
  ErrorInfoResponse:
    allOf:
      - $ref: "#/definitions/InfoResponse"
      - type: "object"
        properties:
          path:
            type: "string"
  JsonPatch:
    description: "A JSONPatch document as defined by RFC 6902"
    required:
      - "op"
      - "path"
    properties:
      op:
        type: "string"
        description: "The operation to be performed"
        enum:
          - "add"
          - "remove"
          - "replace"
          - "move"
          - "copy"
          - "test"
      path:
        type: "string"
        description: "A JSON-Pointer"
      value:
        type: "object"
        description: "The value to be used within the operations."
      from:
        type: "string"
        description: "A string containing a JSON Pointer value."